requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-icons==21.3
fontawesomefree==6.1.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
